Added Quick Logo
Quick logo...

2 variations
blue on red, and red on blue .png
blue on red, and red on blue .svg

This might be a stupid name, I don't know just plowing forward with a mockup

Someone skilled at this needs to reimagine...  this is just candy so the prototype does not look like a totally boring wire frame.

What you got?
LFG